pyproject.toml
[project]
name = "flaskreact"
version = "0.1.0"
description = "a test ml/llm project with flask and react"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.74.1",
    "bs4>=0.0.2",
    "claude-agent-sdk>=0.1.9",
    "ddgs>=9.9.3",
    "dotenv>=0.9.9",
    "duckduckgo-search>=8.1.1",
    "fastmcp>=2.13.1",
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "github-template",
    "langchain>=1.0.2",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.1",
    "langchain-tavily>=0.2.12",
    "langgraph>=1.0.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.3",
    "openai-agents>=0.3.1",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.6.1",
    "pypdf>=6.1.3",
    "pyright>=1.1.407",
    "pytest>=8.4.1",
    "ruff>=0.12.2",
    "scikit-learn>=1.7.0",
    "scikit-learn-stubs>=0.0.3",
    "streamlit>=1.48.0",
    "tool>=0.8.0",
    "torch>=2.7.1",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
python-preference = "system"

[tool.uv.sources]
github_template = { path = "packages/github_template", editable = true }

# ======= ruff設定 =======
[tool.ruff]
line-length = 88
fix = true
indent-width = 4

[tool.ruff.lint]
select = ["I"]  # インポートのソートを有効化

# ======= pytest設定 =======
[tool.pytest.ini_options]
# Pythonパスに追加するディレクトリ
pythonpath = ["src"]
# テストファイルを探索するディレクトリ
testpaths = ["tests"]

markers = [
    "visual: generates plot/image artifacts for manual inspection; skipped unless RUN_VISUAL_TESTS=1 or selected via -m visual",
]

# テストファイルの命名パターン
python_files = ["test_*.py", "*_test.py"]

# テスト関数の命名パターン
python_functions = ["test_*"]

# テストクラスの命名パターン
python_classes = ["Test*"]

# ======= pyright設定 =======
[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "standard"  # "strict"は厳しすぎることも
reportMissingTypeStubs = "warning"
reportUnknownMemberType = "warning"
venvPath = "."
venv = ".venv"